arm loc

+++ include/robots/comp-24/arm.h
@@ -7,7 +7,7 @@ private:
     pros::Rotation *armRot;
     int armState;
     // int targetPosition[3] = {0, 1600, 14000};
-    double targetPosition[4] = {-.5, 10, 26, 140};
+    int targetPosition[4] = {-.5, 10, 26, 140};
     bool intakePullBackFlag;
     bool armFlag;
     double armAngle;
